ml_full_pipeline/requirements.txt
pandas==1.3.0
spacy==3.0.3
pytorch==1.8.1
cudatoolkit==10.2
simpletransformers